+++ app/controllers/about_controller.rb
@@ -1,4 +1,6 @@
 class AboutController < ApplicationController
+  skip_before_action :authenticate_user!
+
   def index
     data = About::IndexData.call
     @about_sections = data.about_sections