+++ app/controllers/application_controller.rb
@@ -1,18 +1,5 @@
 class ApplicationController < ActionController::Base
-  before_action :authenticate_user!, unless: :public_controller?
-
-  PUBLIC_CONTROLLERS = %w[
-    about
-    blackjack
-    blog_posts
-    contact
-    home
-    landscaping
-    projects
-    services
-    s3_proxy
-    videos
-  ].freeze
+  before_action :authenticate_user!
 
   private
 
@@ -21,9 +8,4 @@ class ApplicationController < ActionController::Base
     total_pages = (scope.count / per_page.to_f).ceil
     [scope.offset((page - 1) * per_page).limit(per_page), page, total_pages]
   end
-
-  def public_controller?
-    devise_controller? || PUBLIC_CONTROLLERS.include?(controller_name)
-  end
-
 end