+++ app/controllers/blackjack_controller.rb
@@ -1,4 +1,5 @@
 class BlackjackController < ApplicationController
+  skip_before_action :authenticate_user!
   protect_from_forgery with: :exception
 
   def show