+++ app/controllers/calendar_controller.rb
@@ -1,7 +1,4 @@
 class CalendarController < ApplicationController
-
-  before_action :authenticate_user!
-
   def show
     data = Calendar::ShowData.call
     @tasks = data.tasks