+++ app/controllers/dashboard_controller.rb
@@ -1,6 +1,4 @@
 class DashboardController < ApplicationController
-  before_action :authenticate_user!
-
   def index
     data = Dashboard::IndexData.call(paginator: method(:paginate))