+++ app/controllers/documents_controller.rb
@@ -1,5 +1,4 @@
 class DocumentsController < ApplicationController
-  before_action :authenticate_user!
   before_action :set_document, only: %i[ show destroy ]
 
   # GET /documents or /documents.json