+++ app/controllers/feedbacks_controller.rb
@@ -1,5 +1,4 @@
 class FeedbacksController < ApplicationController
-  before_action :authenticate_user!
   before_action :set_feedback, only: [:edit, :update]
 
   def index