+++ app/controllers/games_controller.rb
@@ -1,6 +1,4 @@
 class GamesController < ApplicationController
-  before_action :authenticate_user!
-  
   def index
     data = Games::IndexData.call(paginator: method(:paginate))
     @games = data.games