+++ app/controllers/goals_controller.rb
@@ -1,5 +1,4 @@
 class GoalsController < ApplicationController
-  before_action :authenticate_user!
   before_action :set_goal, only: %i[ show edit update destroy ]
 
   # GET /goals or /goals.json