+++ app/controllers/home_controller.rb
@@ -1,4 +1,6 @@
 class HomeController < ApplicationController
+  skip_before_action :authenticate_user!
+
   def index
     data = Home::IndexData.call
     @featured_projects = data.featured_projects