+++ app/controllers/ideas_controller.rb
@@ -1,6 +1,4 @@
 class IdeasController < ApplicationController
-  before_action :authenticate_user!
-
   def show
     data = Ideas::ShowData.call(idea_id: params[:id])
     @idea = data.idea