+++ app/controllers/landscaping_controller.rb
@@ -1,4 +1,6 @@
 class LandscapingController < ApplicationController
+  skip_before_action :authenticate_user!
+
   def index
     data = Landscaping::IndexData.call
     @LandscapingJobs = data.jobs