+++ app/controllers/s3_proxy_controller.rb
@@ -1,4 +1,6 @@
 class S3ProxyController < ApplicationController
+  skip_before_action :authenticate_user!
+
   def show
     data = S3Proxy::ShowData.call(key: params[:key], format: params[:format])
     return head :not_found if data.redirect_url.blank?